NavBar init - better layout
